refactor: modularize quickstart shop
